requirements.txt
annotated-types==0.7.0
anyio==4.4.0
assemblyai==0.31.0
audioread==3.0.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.9.0
future==1.0.0
google-api-core==2.19.1
google-api-python-client==2.146.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
lazy-loader==0.4
openai==1.41.0
packaging==24.1
pip==24.2
platformdirs==4.2.2
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
pydub==0.25.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
tqdm==4.66